Winter break Progress
I took a bit of a break from UE4, however this week I tried to make some progress:

So far I have made two functions that will hopefully streamline the socket-finding process, as well as the determining-which-socket-is-closest process. They should make it much easier to add more sockets/ habitat pieces.
(I haven't implemented the new functions in the event graph yet, but I will try and test them out soon)

I still need to work on the sockets themselves for each habitat piece (ex: being able to find all the sockets of a piece, not just one). I also need to work on including the sockets from the user placed pieces in the code, not just the already placed pieces.